refactor: one file for each table
